+++ busy_beaver/factories/event_details.py
