+++ tests/apps/upcoming_events/cards_test.py
@@ -1,5 +1,5 @@
 from busy_beaver.apps.upcoming_events.cards import UpcomingEvent, UpcomingEventList
-from busy_beaver.apps.upcoming_events.factories import EventDetailsFactory
+from busy_beaver.factories.event_details import EventDetailsFactory
 from busy_beaver.toolbox.slack_block_kit import Context, Divider, Section
